clean up packaging and dependencies
