+++ pose_graph_tools_ros/package.xml
@@ -9,13 +9,11 @@
   <license>BSD</license>
 
   <buildtool_depend>ament_cmake</buildtool_depend>
-  <depend>geometry_msgs</depend>
   <depend>interactive_markers</depend>
   <depend>pose_graph_tools</depend>
   <depend>pose_graph_tools_msgs</depend>
   <depend>rclcpp</depend>
   <depend>rclcpp_components</depend>
-  <depend>std_msgs</depend>
   <depend>tf2_eigen</depend>
   <depend>visualization_msgs</depend>
